Refactor the effects placement.
