+++ assets/js/components/ModuleDashboardEffects.js
@@ -1,5 +1,5 @@
 /**
- * ModuleRootComponents component.
+ * ModuleDashboardEffects component.
  *
  * Site Kit by Google, Copyright 2024 Google LLC
  *
